Improve equivalent C of remaining NONMATCHING wrappers
Second-wave pass: semantic verification of all 17 remaining wrapped
functions against the reference asm (full traces for 12, targeted spot
checks for the rest; no semantic bugs found). Banked improvements to the
equivalent C of sub_08025650 (table-base hoisting), sub_0811BEBC
(u8->u32 locals, single-swap residual) and sub_0802E78C (register pin +
per-block clamp pointers). make compare passes (katam.gba: OK).

+++ src/object_spawner.c
@@ -139,10 +139,10 @@ static NAKED void sub_0811BEBC(struct Task *task) {
 #else
 static void sub_0811BEBC(struct Task *task) {
     struct ObjectSpawner *x = TaskGetStructPtr(task);
-    u8 playerId = x->obj2.base.unk56;
-    u8 a = x->obj2.object->unk2;
-    u8 b = x->obj2.object->unk3;
-    u8 c = gCurLevelInfo[playerId].unk65E;
+    u32 playerId = x->obj2.base.unk56;
+    u32 a = x->obj2.object->unk2;
+    u32 b = x->obj2.object->unk3;
+    u32 c = gCurLevelInfo[playerId].unk65E;
 
     ObjectDestroy(task);
     if (a != 0 || playerId != 0xFF) {
